cleanup TypeSafeRoundTripSerialisationExample
